backend/requirements.txt
Django==3.1
psycopg2
gunicorn
django-heroku
parameterized
djangorestframework
django-cors-headers